Update App to build Exchange from py script, as well as Read from .csv for ticker names.  Currently assign values, need to query real values

+++ src/Stock.cpp
@@ -3,7 +3,7 @@
 #include "Stock.h"
 
 
-Stock::Stock(std::string tick, int val):
+Stock::Stock(std::string tick, float val):
 ticker(tick),
 value(val),
 lastValue(0.0f),
